refactor: touch-listener to own mod; but re-exported for consumers
